+++ apps/zbugs/src/components/emoji-data-source.ts
@@ -1,0 +1,3 @@
+import emojiDataSource from 'emoji-picker-element-data/en/emojibase/data.json?url';
+
+export {emojiDataSource};
