+++ apps/zbugs/src/pages/issue/comment.tsx
@@ -8,13 +8,13 @@ import {CanEdit} from '../../components/can-edit.js';
 import {Confirm} from '../../components/confirm.js';
 import {EmojiPanel} from '../../components/emoji-panel.js';
 import {Link} from '../../components/link.js';
-import Markdown from '../../components/markdown.js';
-import RelativeTime from '../../components/relative-time.js';
+import {Markdown} from '../../components/markdown.js';
+import {RelativeTime} from '../../components/relative-time.js';
 import {type Emoji} from '../../emoji-utils.js';
 import {useHash} from '../../hooks/use-hash.js';
 import {useLogin} from '../../hooks/use-login.js';
 import {useZero} from '../../hooks/use-zero.js';
-import CommentComposer from './comment-composer.js';
+import {CommentComposer} from './comment-composer.js';
 import style from './comment.module.css';
 
 type Props = {
@@ -33,98 +33,99 @@ type Props = {
   highlight?: boolean | undefined;
 };
 
-const Comment = memo(({id, issueID, comment, height, highlight}: Props) => {
-  const z = useZero();
-  const [editing, setEditing] = useState(false);
-  const login = useLogin();
-  const [deleteConfirmationShown, setDeleteConfirmationShown] = useState(false);
+export const Comment = memo(
+  ({id, issueID, comment, height, highlight}: Props) => {
+    const z = useZero();
+    const [editing, setEditing] = useState(false);
+    const login = useLogin();
+    const [deleteConfirmationShown, setDeleteConfirmationShown] =
+      useState(false);
 
-  const hash = useHash();
-  const permalink = comment && makePermalink(comment);
-  const isPermalinked = highlight || hash === permalink;
+    const hash = useHash();
+    const permalink = comment && makePermalink(comment);
+    const isPermalinked = highlight || hash === permalink;
 
-  const edit = () => setEditing(true);
-  const remove = () => z.mutate.comment.delete({id});
+    const edit = () => setEditing(true);
+    const remove = () => z.mutate.comment.delete({id});
 
-  if (!comment) {
-    return <div style={{height}}></div>;
-  }
-  return (
-    <div
-      className={classNames({
-        [style.commentItem]: true,
-        [style.authorComment]:
-          comment.creatorID == login.loginState?.decoded.sub,
-        [style.permalinked]: isPermalinked,
-      })}
-    >
-      <p className={style.commentAuthor}>
-        <img
-          src={comment.creator?.avatar}
-          style={{
-            width: '2rem',
-            height: '2rem',
-            borderRadius: '50%',
-            display: 'inline-block',
-            marginRight: '0.3rem',
-          }}
-          alt={comment.creator?.name ?? undefined}
-        />{' '}
-        {comment.creator?.login}
-      </p>
-      <span id={permalink} className={style.commentTimestamp}>
-        <Link href={`#${permalink}`}>
-          <RelativeTime timestamp={comment.created} />
-        </Link>
-      </span>
-      {editing ? (
-        <CommentComposer
-          id={id}
-          body={comment.body}
-          issueID={issueID}
-          onDone={() => setEditing(false)}
-        />
-      ) : (
-        <>
-          <div className="markdown-container">
-            <Markdown>{comment.body}</Markdown>
-          </div>
-          <EmojiPanel
+    if (!comment) {
+      return <div style={{height}}></div>;
+    }
+    return (
+      <div
+        className={classNames({
+          [style.commentItem]: true,
+          [style.authorComment]:
+            comment.creatorID == login.loginState?.decoded.sub,
+          [style.permalinked]: isPermalinked,
+        })}
+      >
+        <p className={style.commentAuthor}>
+          <img
+            src={comment.creator?.avatar}
+            style={{
+              width: '2rem',
+              height: '2rem',
+              borderRadius: '50%',
+              display: 'inline-block',
+              marginRight: '0.3rem',
+            }}
+            alt={comment.creator?.name ?? undefined}
+          />{' '}
+          {comment.creator?.login}
+        </p>
+        <span id={permalink} className={style.commentTimestamp}>
+          <Link href={`#${permalink}`}>
+            <RelativeTime timestamp={comment.created} />
+          </Link>
+        </span>
+        {editing ? (
+          <CommentComposer
+            id={id}
+            body={comment.body}
             issueID={issueID}
-            commentID={comment.id}
-            emojis={comment.emoji}
+            onDone={() => setEditing(false)}
           />
-        </>
-      )}
-      {editing ? null : (
-        <CanEdit ownerID={comment.creatorID}>
-          <div className={style.commentActions}>
-            <Button eventName="Edit comment" onAction={edit}>
-              Edit
-            </Button>
-            <Button
-              eventName="Delete comment"
-              onAction={() => setDeleteConfirmationShown(true)}
-            >
-              Delete
-            </Button>
-          </div>
-        </CanEdit>
-      )}
-      <Confirm
-        title="Delete Comment"
-        text="Deleting a comment is permanent. Are you sure you want to delete this comment?"
-        okButtonLabel="Delete"
-        isOpen={deleteConfirmationShown}
-        onClose={b => {
-          if (b) {
-            remove();
-          }
-          setDeleteConfirmationShown(false);
-        }}
-      />
-    </div>
-  );
-});
-
-export {Comment as default};
+        ) : (
+          <>
+            <div className="markdown-container">
+              <Markdown>{comment.body}</Markdown>
+            </div>
+            <EmojiPanel
+              issueID={issueID}
+              commentID={comment.id}
+              emojis={comment.emoji}
+            />
+          </>
+        )}
+        {editing ? null : (
+          <CanEdit ownerID={comment.creatorID}>
+            <div className={style.commentActions}>
+              <Button eventName="Edit comment" onAction={edit}>
+                Edit
+              </Button>
+              <Button
+                eventName="Delete comment"
+                onAction={() => setDeleteConfirmationShown(true)}
+              >
+                Delete
+              </Button>
+            </div>
+          </CanEdit>
+        )}
+        <Confirm
+          title="Delete Comment"
+          text="Deleting a comment is permanent. Are you sure you want to delete this comment?"
+          okButtonLabel="Delete"
+          isOpen={deleteConfirmationShown}
+          onClose={b => {
+            if (b) {
+              remove();
+            }
+            setDeleteConfirmationShown(false);
+          }}
+        />
+      </div>
+    );
+  },
+);
