+++ apps/zbugs/src/pages/issue/issue-page.tsx
@@ -30,16 +30,16 @@ import {CanEdit} from '../../components/can-edit.js';
 import {Combobox} from '../../components/combobox.js';
 import {Confirm} from '../../components/confirm.js';
 import {EmojiPanel} from '../../components/emoji-panel.js';
-import {useEmojiDataSourcePreload} from '../../components/emoji-picker.js';
-import LabelPicker from '../../components/label-picker.js';
+import {LabelPicker} from '../../components/label-picker.js';
 import {Link} from '../../components/link.js';
-import Markdown from '../../components/markdown.js';
-import RelativeTime from '../../components/relative-time.js';
-import UserPicker from '../../components/user-picker.js';
+import {Markdown} from '../../components/markdown.js';
+import {RelativeTime} from '../../components/relative-time.js';
+import {UserPicker} from '../../components/user-picker.js';
 import {type Emoji} from '../../emoji-utils.js';
 import {useCanEdit} from '../../hooks/use-can-edit.js';
 import {useDocumentHasFocus} from '../../hooks/use-document-has-focus.js';
-import useIsScrolling from '../../hooks/use-is-scrolling.js';
+import {useEmojiDataSourcePreload} from '../../hooks/use-emoji-data-source-preload.js';
+import {useIsScrolling} from '../../hooks/use-is-scrolling.js';
 import {useKeypress} from '../../hooks/use-keypress.js';
 import {useLogin} from '../../hooks/use-login.js';
 import {useZero} from '../../hooks/use-zero.js';
@@ -47,8 +47,8 @@ import {LRUCache} from '../../lru-cache.js';
 import {recordPageLoad} from '../../page-load-stats.js';
 import {links, type ListContext, type ZbugsHistoryState} from '../../routes.js';
 import {preload} from '../../zero-setup.js';
-import CommentComposer from './comment-composer.js';
-import Comment from './comment.js';
+import {CommentComposer} from './comment-composer.js';
+import {Comment} from './comment.js';
 import {isCtrlEnter} from './is-ctrl-enter.js';
 
 const emojiToastShowDuration = 3_000;
