+++ apps/zbugs/src/pages/list/list-page.tsx
@@ -16,18 +16,18 @@ import {useSearch} from 'wouter';
 import {navigate} from 'wouter/use-browser-location';
 import {Button} from '../../components/button.js';
 import {Filter, type Selection} from '../../components/filter.js';
-import IssueLink from '../../components/issue-link.js';
+import {IssueLink} from '../../components/issue-link.js';
 import {Link} from '../../components/link.js';
-import RelativeTime from '../../components/relative-time.js';
+import {RelativeTime} from '../../components/relative-time.js';
 import {useClickOutside} from '../../hooks/use-click-outside.js';
 import {useElementSize} from '../../hooks/use-element-size.js';
 import {useKeypress} from '../../hooks/use-keypress.js';
 import {useLogin} from '../../hooks/use-login.js';
 import {useZero} from '../../hooks/use-zero.js';
+import {recordPageLoad} from '../../page-load-stats.js';
 import {mark} from '../../perf-log.js';
 import type {ListContext} from '../../routes.js';
 import {preload} from '../../zero-setup.js';
-import {recordPageLoad} from '../../page-load-stats.js';
 
 let firstRowRendered = false;
 const itemSize = 56;
